HFM: Initial Commitment
This is my personal version of the HFM Mod, incorporating elements of
PDM, that awesome re-work of political parties that I haven't been able
to find again, and more. It is not intended to be shared, but you're
welcome to download it if you wish.

I am doing this in order to update the Mod without erasing my stuff.